# CRITICAL: PLUMCP Security Hardening - Complete Vulnerability Remediation & Performance Optimization