--- a/tensordemo/fileSystemService_pr5_fixed.ts
+++ b/tensordemo/fileSystemService_pr5_fixed.ts
@@ -1,0 +1,82 @@
+// Fixed version of fileSystemService.ts for PR5 with all bot feedback addressed
+
+// ... existing imports ...
+import { createHash } from 'node:crypto';
+
+// ... existing code ...
+
+// Fixed AgentType enum with CODE_GENERATION added
+export enum AgentType {
+  CONFLICT_RESOLUTION = 'conflict_resolution',
+  CONSISTENCY_MAINTENANCE = 'consistency_maintenance',
+  RECOVERY_SPECIALIST = 'recovery_specialist',
+  OPTIMIZATION_ENGINE = 'optimization_engine',
+  SECURITY_GUARDIAN = 'security_guardian',
+  ANALYSIS_INSIGHT = 'analysis_insight',
+  CODE_GENERATION = 'code_generation' // Added for unique agent dispatching
+}
+
+// Fixed hashContent method with SHA-256
+private hashContent(content: string): string {
+  // Using cryptographic hash to prevent collisions and ensure data integrity
+  return createHash('sha256').update(content).digest('hex');
+}
+
+// Fixed CodeGenerationAgent with proper unique type
+class CodeGenerationAgent implements VFSAgent {
+  readonly type = AgentType.CODE_GENERATION; // Using unique type to prevent agent collisions
+  readonly name = 'Code Generation Agent';
+  readonly priority = AgentPriority.MEDIUM;
+  readonly triggers = [AgentTrigger.ON_DEMAND];
+
+  // ... rest of implementation ...
+}
+
+// Fixed conflict resolution MERGE case to throw error instead of misleading overwrite
+private async handleConflictResolution(filePath: string, resolution: ConflictResolution): Promise<void> {
+  switch (resolution) {
+    case ConflictResolution.OVERWRITE:
+      // Overwrite with new content
+      if (this.config.enableLogging) {
+        console.log(`[VFS] Resolving conflict by overwriting ${filePath}`);
+      }
+      break;
+
+    case ConflictResolution.KEEP_EXISTING:
+      // Keep existing content, discard new
+      if (this.config.enableLogging) {
+        console.log(`[VFS] Resolving conflict by keeping existing ${filePath}`);
+      }
+      break;
+
+    case ConflictResolution.MERGE:
+      // True merge not yet implemented - throw error to prevent data loss
+      throw new Error(`File conflict: ${filePath} requires a merge, but auto-merging is not yet implemented. Please resolve manually.`);
+
+    case ConflictResolution.MANUAL:
+      // Require manual intervention
+      throw new Error(`File conflict: ${filePath} requires manual resolution.`);
+  }
+}
+
+// Fixed GuidanceSystem lazy loading with proper ESM dynamic import
+private async initializeGuidanceSystem(): Promise<void> {
+  try {
+    // Use proper ESM dynamic import instead of require()
+    const { GuidanceSystem } = await import('../utils/guidance.js');
+    this.guidanceSystem = new GuidanceSystem();
+  } catch (error) {
+    // Guidance system not available, continue without it
+    if (this.config.enableLogging) {
+      console.warn('[VFS] Guidance system not available for conflict resolution');
+    }
+  }
+}
+
+// Update constructor to use the new initialization method
+constructor() {
+  // Initialize guidance system asynchronously
+  this.initializeGuidanceSystem();
+}
+
+// ... rest of existing code ...