--- a/tensordemo/pr1_description.md
+++ b/tensordemo/pr1_description.md
@@ -1,5 +1,7 @@
 🚀 Build Pipeline: 100% Success Rate & Advanced Caching
 
+Fixes #8607 - Fix Critical Build and Test Issues
+
 Enhanced file system operations with intelligent caching and conflict resolution for superior reliability.
 
 ✅ 100% build success rate (eliminated all failures)