--- a/tensordemo/pr2_description.md
+++ b/tensordemo/pr2_description.md
@@ -1,0 +1,20 @@
+⚡ Enhanced Edit Tool: Safe Replacements & Conflict Resolution
+
+Advanced editing capabilities with comprehensive error handling and safety features.
+
+✅ Enhanced edit operations with safety checks
+✅ Conflict resolution integration
+✅ Improved error reporting and debugging
+✅ Better validation and correctness checking
+✅ Safe handling of edge cases and special characters
+
+Dependencies: comment-json, @joshua.litt/get-ripgrep
+Files: packages/core/src/tools/edit.ts
+
+Additional Improvements:
+- Safe string replacement with $ sequence handling
+- Virtual File System integration
+- Enhanced debugging information
+- Comprehensive validation system
+
+Ready to merge! ⚡