--- a/tensordemo/pr2_description.md
+++ b/tensordemo/pr2_description.md
@@ -1,5 +1,7 @@
 ⚡ Enhanced Edit Tool: Safe Replacements & Conflict Resolution
 
+Fixes #8607 - Fix Critical Build and Test Issues
+
 Advanced editing capabilities with comprehensive error handling and safety features.
 
 ✅ Enhanced edit operations with safety checks