--- a/tensordemo/pr3_description.md
+++ b/tensordemo/pr3_description.md
@@ -1,0 +1,21 @@
+📈 Advanced Code Analysis: Security & Quality Intelligence
+
+Comprehensive code analysis system with security vulnerability detection and quality metrics.
+
+✅ Advanced code pattern analysis and classification
+✅ Security vulnerability detection (injection, XSS, weak crypto)
+✅ Quality metrics (maintainability index, cyclomatic complexity)
+✅ Improvement suggestions with refactoring recommendations
+✅ Alternative implementation strategies with scoring
+
+Dependencies: comment-json, @joshua.litt/get-ripgrep
+Files: packages/core/src/utils/guidance.ts
+
+Additional Improvements:
+- Multi-paradigm code analysis
+- Domain context detection
+- Comprehensive security analysis
+- Code quality assessment
+- Intelligent naming convention system
+
+Ready to merge! 📈