--- a/tensordemo/pr3_description.md
+++ b/tensordemo/pr3_description.md
@@ -1,5 +1,7 @@
 📈 Advanced Code Analysis: Security & Quality Intelligence
 
+Fixes #8607 - Fix Critical Build and Test Issues
+
 Comprehensive code analysis system with security vulnerability detection and quality metrics.
 
 ✅ Advanced code pattern analysis and classification