--- a/tensordemo/pr4_description.md
+++ b/tensordemo/pr4_description.md
@@ -1,0 +1,21 @@
+📊 Performance Monitoring: Real-Time Analytics & Health Dashboard
+
+Real-time system monitoring with predictive analytics and comprehensive health tracking.
+
+✅ Real-time performance monitoring and metrics collection
+✅ Predictive analytics with trend analysis and forecasting
+✅ Anomaly detection with ML-based alerting
+✅ Comprehensive health dashboard with WebSocket updates
+✅ System performance optimization recommendations
+
+Dependencies: comment-json, @joshua.litt/get-ripgrep
+Files: packages/core/src/services/fileSystemService.ts, packages/core/src/tools/edit.ts
+
+Additional Improvements:
+- Multi-dimensional health monitoring
+- Intelligent alerting and escalation
+- Performance regression detection
+- Automated system optimization
+- Historical trend analysis
+
+Ready to merge! 📊