--- a/tensordemo/pr4_description.md
+++ b/tensordemo/pr4_description.md
@@ -1,5 +1,7 @@
 📊 Performance Monitoring: Real-Time Analytics & Health Dashboard
 
+Fixes #8607 - Fix Critical Build and Test Issues
+
 Real-time system monitoring with predictive analytics and comprehensive health tracking.
 
 ✅ Real-time performance monitoring and metrics collection