--- a/tensordemo/pr5_description.md
+++ b/tensordemo/pr5_description.md
@@ -1,5 +1,7 @@
 🎯 System Enhancement: Comprehensive Reliability Framework
 
+Fixes #8607 - Fix Critical Build and Test Issues
+
 Complete system reliability framework with advanced monitoring, testing, and deployment automation.
 
 ✅ Complete reliability framework implementation