… packages/core/src/services/fileSystemService.ts
… packages/core/src/utils/guidance.ts